Refactor getSelectFieldData and getCategories to streamline imports and improve clarity
