+++ lib/partners.ts
@@ -2,7 +2,6 @@ import { COLLECTIONS, INVALIDATE_INTERVAL } from '@/constants'
 import { db } from '@/lib'
 import { collection, getDocs, limit, orderBy, query, startAfter, where } from 'firebase/firestore'
 import { unstable_cache } from 'next/cache'
-import { getCollection } from './api'
 
 export const getPartners = async ({
   order,
@@ -84,19 +83,5 @@ export const getPartners = async ({
       },
       data: [],
     }
-  }
-}
-
-export const getCategories = async (locale: LocalesType = 'pt_br') => {
-  try {
-    const result = await getCollection(COLLECTIONS.CATEGORIES, { locale })
-    return {
-      ...result,
-      //@ts-expect-error title type is defined as string, so it could not use a locale as key as an object. Check this.
-      data: (result.data ?? []).sort((a, b) => a.title[locale].localeCompare(b.title[locale])),
-    }
-  } catch (e) {
-    console.error(e)
-    throw Error
   }
 }
